Delete LB that gone into ERROR state (#2341)
This commit implements deleting the LB when it goes into ERROR state on
creation. This will allow us to be less prone to transient Octavia
issues.

A possible improvement could be to extend this behavior to all cases
when we wait for the LB to become ACTIVE. The problem with that is that
attempt to implement that complicates the code a lot, so let's start
with this first step.