Remove enforcement of IPv6 LB as internal
OpenStack does not provide NAT with IPv6, so
it has no support for Floating IP with IPv6.
But, IPv6 that uses GUAs don't require NAT
to access the outside world, so IPv6 can be
rechable without Floating IPs making the
enforcement of IPv6 LB as internal not necessary.
As a result, IPv6 load-balancers are allowed to
be shared between Services.