Remove enforcement of IPv6 LB as internal (#2557) (#2566)
In OpenStack IPv6 that uses GUAs don't require NAT
to access the outside world, so IPv6 can be
rechable without Floating IPs, which makes the
enforcement of IPv6 LB as internal in CPO not necessary.
This commit removes this enforcement, which results in
IPv6 load-balancers being allowed to be shared between
Services. Also, now it's possible to make the load-balancer
use the IPv6 stateful address defined in the loadBalancerIP
of the Service.