Allow changing cluster-name on existing deployments (#2568)
It's a common issue that clusters are deployed with the default
`--cluster-name=kubernetes` and later on it's discovered that next
deployments on the same cloud will have conflicts when trying to manage
LBs of the same namespace and name.

This commit aims at allowing to change the cluster-name on a running
environment and handling all the renames of the LB resources and their
tags.

Co-authored-by: Michał Dulko <mdulko@redhat.com>